chore: adjust imports and folders
