+++ apps/ui/src/modules/marketplace/components/banner/root/index.tsx
@@ -1,0 +1,15 @@
+import { DetailsMenu } from './detailsMenu';
+import { ShareMenu } from './shareMenu';
+import { StatBox } from './statBox';
+import { CollectionsContent } from './collectionsContent';
+import { CollectionContent } from './collectionContent';
+import { SocialActionsMenu } from './socialActionsMenu';
+
+export const BannerRoot = {
+  CollectionsContent: CollectionsContent,
+  CollectionContent: CollectionContent,
+  StatBox: StatBox,
+  ShareMenu: ShareMenu,
+  DetailsMenu: DetailsMenu,
+  SocialActionsMenu: SocialActionsMenu,
+};
