+++ apps/ui/src/modules/marketplace/components/banner/root/shareMenu.tsx
@@ -8,10 +8,10 @@ import {
   Icon,
   Text,
 } from '@chakra-ui/react';
-import { ShareIcon2 } from '../icons/shareIcon2';
-import { TwitterIcon } from '../icons/twitterIcon';
 import { DiscordIcon } from '@/components/icons/discordIcon';
-import { CopyIcon2 } from '../icons';
+import { CopyIcon2 } from '../../icons';
+import { ShareIcon2 } from '../../icons/shareIcon2';
+import { TwitterIcon } from '../../icons/twitterIcon';
 
 const menuItems = [
   {
