+++ apps/ui/src/modules/marketplace/components/index.ts
@@ -1,4 +1,4 @@
-export * from './banner';
+export * from './banner/collectionBanner';
 export * from './marketplaceHeader';
 export * from './orderList';
 export { default as SearchBar } from './searchBar';
